organized file structure and added styles.css
