+++ src/index.html
@@ -3,6 +3,7 @@
   <head>
     <meta charset="UTF-8" />
     <title>KMon</title>
+    <link rel="stylesheet" href="styles/styles.css">
   </head>
   <body>
     <div id ="root">
