+++ src/pages/Consumers.jsx
@@ -4,7 +4,7 @@ const Consumers = () => {
 
   return (
     <div>
-      <h1>Consumers</h1>
+      <h1>This is the Comsumers Page</h1>
     </div>
   );
 };
